[closes #33]: create 2 tests for color palette feature before its implementation (TDD)
Both tests are failing right now and are waiting for the correct code to be
implemented in the ArtPieceDetails component.
So this is basically Test-Driven-Development.
One tests checks if there a list item with the hex code of one of the colors
of the rendered piece.
The other test checks if there a actually all 5 colors rendered as list items.